requirements.txt
# Minimal requirements for hackathon demo
# Install these one by one if having issues

# Core dependencies
numpy>=1.21.0
langchain>=0.1.0
huggingface-hub>=0.16.0
sentence-transformers>=2.2.0

# Vector search (CPU version)
faiss-cpu>=1.7.0

# Document processing
PyMuPDF>=1.23.0

# Basic utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
loguru>=0.7.0

# Tokenization
tiktoken>=0.5.0